update semester block style
